Refactoring to use modular world models rather than agent models.
Both planning agents and replanning agent models can be specified with a 
unified interface, along with agents with changing goals and 
environments with stochastic dynamics.

The new abstraction is a `world_step`, which consists of a `goal_step` 
that possibly changes the goal / keeps it the same, passes that to a 
`plan_step` which steps the plan forward in time, possibly replanning if 
a plan hasn't been made yet. This outputs the current action, which is 
passed to `env_step` which advances the state of the environment, and 
then the new state is passed to `obs_step` , which adds (optionally 
history-dependent) observation noise.